pyproject.toml
[project]
name = "low-shot-color-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "onnxruntime>=1.24.4",
    "pillow>=12.2.0",
]
